+++ lib/routes/index.js
@@ -16,6 +16,7 @@ module.exports = function(crowi, app) {
   var revision = require('./revision')(crowi, app)
   var search = require('./search')(crowi, app)
   var share = require('./share')(crowi, app)
+  var access = require('./access')(crowi, app)
   var loginRequired = middleware.loginRequired
   var fileAccessRightOrLoginRequired = middleware.fileAccessRightOrLoginRequired(crowi, app)
   var accessTokenParser = middleware.accessTokenParser(crowi, app)
@@ -250,6 +251,8 @@ module.exports = function(crowi, app) {
     share.api.setSecretKeyword,
   )
   app.post('/_api/shares/secretKeyword.check', csrf, share.api.checkSecretKeyword)
+
+  app.get('/_api/accesses.list', accessTokenParser, loginRequired(crowi, app), access.api.list)
 
   app.post('/_/edit', form.revision, loginRequired(crowi, app), csrf, page.pageEdit)
   app.get('/trash/$', loginRequired(crowi, app), page.deletedPageListShow)
