+++ package.json
@@ -93,6 +93,7 @@
     "passport": "^0.4.0",
     "passport-github": "^1.1.0",
     "plantuml-encoder": "~1.2.4",
+    "platform": "^1.3.5",
     "react": "~15.5.0",
     "react-bootstrap": "~0.30.10",
     "react-dom": "~15.5.0",
